+++ crates/client/eth/src/config.rs
@@ -1,4 +1,3 @@
-
 use std::fs::File;
 use std::path::PathBuf;
 use alloy::primitives::{
@@ -12,10 +11,6 @@ use crate::error::Error;
 pub const DEFAULT_RPC_ENDPOINT: &str = "http://127.0.0.1:8545";
 /// Default Anvil chain ID
 pub const DEFAULT_CHAIN_ID: u64 = 31337;
-/// Default private key derived from starting Anvil as follows:
-/// anvil -b 5 --config-out $BUILD_DIR/anvil.json
-/// PRE_PRIVATE=$(jq -r '.private_keys[0]' $BUILD_DIR/anvil.json)
-pub const DEFAULT_PRIVATE_KEY: &str = "ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80";
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 pub struct EthereumClientConfig {
@@ -48,10 +43,6 @@ fn default_rpc_endpoint() -> String {
 
 fn default_chain_id() -> u64 {
     DEFAULT_CHAIN_ID
-}
-
-fn default_private_key() -> String {
-    DEFAULT_PRIVATE_KEY.to_string()
 }
 
 impl Default for HttpProviderConfig {
