+++ crates/client/eth/src/error.rs
@@ -8,4 +8,5 @@ pub enum Error {
     ConfigReadFromFile(#[from] std::io::Error),
     #[error("Failed to decode from JSON: {0}")]
     ConfigDecodeFromJson(#[from] serde_json::Error),
+
 }
