+++ crates/client/sync/Cargo.toml
@@ -18,7 +18,6 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 
-dc-eth = { path = "./src/eth" }
 # Deoxys
 dc-db = { workspace = true }
 dc-eth = { workspace = true }
@@ -31,6 +30,7 @@ dp-receipt = { workspace = true }
 dp-state-update = { workspace = true }
 dp-transactions = { workspace = true }
 dp-utils = { workspace = true }
+dc-eth = {workspace  = true}
 
 # Starknet
 bonsai-trie = { workspace = true }
