--- crates/client/sync/src/eth/Cargo.toml
@@ -1,70 +1,0 @@
-[package]
-authors.workspace = true
-description = "This crate is responsible to handle l1 communication"
-edition = "2021"
-homepage = "https://github.com/KasarLabs/deoxys"
-license = "MIT"
-name = "dc-eth"
-publish = false
-repository = "https://github.com/KasarLabs/deoxys"
-version = "0.1.0"
-
-[features]
-default = []
-m = ["dep:rodio"]
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-
-# Deoxys
-dc-db = { workspace = true }
-dc-metrics = { workspace = true }
-dc-telemetry = { workspace = true }
-dp-block = { workspace = true }
-dp-class = { workspace = true }
-dp-convert = { workspace = true }
-dp-receipt = { workspace = true }
-dp-transactions = { workspace = true }
-dp-utils = { workspace = true }
-
-# Starknet
-blockifier = { workspace = true, features = ["testing"] }
-bonsai-trie = { workspace = true }
-starknet-core = { workspace = true }
-starknet-providers = { workspace = true }
-starknet-types-core = { workspace = true }
-starknet_api = { workspace = true }
-
-
-# Other
-anyhow = "1.0.75"
-bitvec = { workspace = true }
-rustc-hex = { workspace = true }
-ethers = { workspace = true }
-futures = { workspace = true, default-features = true }
-hex = { workspace = true }
-indexmap = { workspace = true }
-itertools = { workspace = true }
-log = { workspace = true }
-num-traits = { workspace = true }
-primitive-types = { workspace = true }
-rand = { workspace = true }
-rayon = { workspace = true }
-reqwest = { workspace = true }
-rodio = { version = "0.17", optional = true }
-serde = { workspace = true, default-features = true }
-serde_json = "1"
-alloy = { version = "0.2.0", features = ["full"] }
-thiserror.workspace = true
-tokio = { workspace = true, features = [
-    "macros",
-    "parking_lot",
-    "test-util",
-    "signal",
-] }
-url = { workspace = true }
-
-[dev-dependencies]
-# test_utils = { path = "./test_utils" }
